Fix: semester을 number -\> string으로 타입 수정 , 테이블 데이터들이 모두 select 되지 않는  버그 수정(#4)
